backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
openai==1.3.0
anthropic==0.7.0
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
prometheus-client==0.19.0